--- a/src/jit/utils.cpp
+++ b/src/jit/utils.cpp
@@ -1897,8 +1897,8 @@ double FloatingPointUtils::round(double x)
 
     // This is based on the 'Berkeley SoftFloat Release 3e' algorithm
 
-    uint64_t bits = *reinterpret_cast<uint64_t*>(&x);
-    int32_t exponent = (int32_t)(bits >> 52) & 0x07FF;
+    uint64_t bits     = *reinterpret_cast<uint64_t*>(&x);
+    int32_t  exponent = (int32_t)(bits >> 52) & 0x07FF;
 
     if (exponent <= 0x03FE)
     {
@@ -1907,7 +1907,7 @@ double FloatingPointUtils::round(double x)
             // Exactly +/- zero should return the original value
             return x;
         }
-        
+
         // Any value less than or equal to 0.5 will always round to exactly zero
         // and any value greater than 0.5 will always round to exactly one. However,
         // we need to preserve the original sign for IEEE compliance.
@@ -1930,7 +1930,7 @@ double FloatingPointUtils::round(double x)
     // Determine the last bit that represents the integral portion of the value
     // and the bits representing the fractional portion
 
-    uint64_t lastBitMask = UI64(1) << (0x0433 - exponent);
+    uint64_t lastBitMask   = UI64(1) << (0x0433 - exponent);
     uint64_t roundBitsMask = lastBitMask - 1;
 
     // Increment the first fractional bit, which represents the midpoint between
@@ -1979,8 +1979,8 @@ float FloatingPointUtils::round(float x)
 
     // This is based on the 'Berkeley SoftFloat Release 3e' algorithm
 
-    uint32_t bits = *reinterpret_cast<uint32_t*>(&x);
-    int32_t exponent = (int32_t)(bits >> 23) & 0xFF;
+    uint32_t bits     = *reinterpret_cast<uint32_t*>(&x);
+    int32_t  exponent = (int32_t)(bits >> 23) & 0xFF;
 
     if (exponent <= 0x7E)
     {
@@ -2012,7 +2012,7 @@ float FloatingPointUtils::round(float x)
     // Determine the last bit that represents the integral portion of the value
     // and the bits representing the fractional portion
 
-    uint32_t lastBitMask = 1U << (0x96 - exponent);
+    uint32_t lastBitMask   = 1U << (0x96 - exponent);
     uint32_t roundBitsMask = lastBitMask - 1;
 
     // Increment the first fractional bit, which represents the midpoint between