This resolves https://github.com/dotnet/coreclr/issues/25857 and also brings in a nice little perf increase for the managed implementation.

### Before Intrinsic

| Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|  Round | 4.169 us | 0.0134 us | 0.0119 us | 4.168 us | 4.153 us | 4.187 us |           - |           - |           - |                   - |
|  Round | 4.183 us | 0.0118 us | 0.0111 us | 4.181 us | 4.165 us | 4.201 us |           - |           - |           - |                   - |

### After Intrinsic

| Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|  Round | 4.150 us | 0.0064 us | 0.0060 us | 4.151 us | 4.141 us | 4.161 us |           - |           - |           - |                   - |
|  Round | 4.164 us | 0.0107 us | 0.0100 us | 4.166 us | 4.146 us | 4.179 us |           - |           - |           - |                   - |

### Before Managed

| Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|  Round | 32.12 us | 0.0903 us | 0.0845 us | 32.14 us | 31.94 us | 32.24 us |           - |           - |           - |                   - |
|  Round | 25.03 us | 0.1173 us | 0.1040 us | 25.02 us | 24.81 us | 25.20 us |           - |           - |           - |                   - |

### After Managed

| Method |     Mean |     Error |    StdDev |   Median |      Min |      Max | Gen 0/1k Op | Gen 1/1k Op | Gen 2/1k Op | Allocated Memory/Op |
|------- |---------:|----------:|----------:|---------:|---------:|---------:|------------:|------------:|------------:|--------------------:|
|  Round | 17.29 us | 0.0432 us | 0.0404 us | 17.27 us | 17.23 us | 17.36 us |           - |           - |           - |                   - |
|  Round | 15.77 us | 0.0150 us | 0.0140 us | 15.77 us | 15.75 us | 15.80 us |           - |           - |           - |                   - |