@tannergooding 2019-07-26
CC. @danmosemsft.

We want to backport this to nc3.0 given it impacts determinism under tiering, correct?