@danmosemsft 2019-07-26
> We want to backport this to nc3.0 given it impacts determinism under tiering, correct?

The impact certainly meets the bar - numeric behavior changing behavior spontanously is a recipe for heisenbugs. 

But the change seems large and potentially risky. Thoughts about that? Do we have enough coverage on different platforms? 

Is there a way to run all CoreFX tests without tiering (ie compare a run with all "optimized JIT" and another run with all "quick JIT") to see whether any other behavior is different in the "no tiering" regime?

cc @jkotas @kouvel