@danmosemsft 2019-07-26
Is there a more targeted, safer fix that could be done for 3.0 only? I don't care about the perf improvement for 3.0