@danmosemsft 2019-07-26
Are there more tests we can usefully add (on top of testing with and without tiering)? That could also reduce risk