@tannergooding 2019-07-26
> Is there a more targeted, safer fix that could be done for 3.0 only? I don't care about the perf improvement for 3.0

This is the targeted fix and is using a known good/existing implementation. It just happens to bring in a perf benefit at the same time.

> Are there more tests we can usefully add (on top of testing with and without tiering)? That could also reduce risk

Tests have to be added to CoreFX, and I am working on getting those up.