@tannergooding 2019-07-26
> Also can you briefly educate me how tiering caused htis?

Under Tier 0, the JIT will just use the pre-jitted `System.Math.Round` and `System.MathF.Round` implementations. These implementations have a bug and do not always return the IEEE compliant result.

After a certain number of calls (30), the JIT will rejit these methods for the current hardware. On machines with SSE4.1 support (most machines in the past 12 years), the JIT will treat these methods as "intrinsic" and just directly emit the `roundsd` and `roundss` instruction instead of compiling the C# implementation. These instructions are IEEE compliant.

This difference can cause cold vs hot methods to return different results for some inputs (for example, as is the case with `0.499999969f`).