@tannergooding 2019-07-26
This PR just updates the implementation to use a known good/correct implementation which should always return the correct result (and therefore the behavior should always match between the various tiers).