@EgorBo 2019-07-26
I found the problematic value using this script:
```csharp
using System;
using System.Runtime.CompilerServices;

class Program
{
    static void Main()
    {
        Random rand = new Random();
        for (int i = 0; i < int.MaxValue; i++)
        {
            double rd = rand.NextDouble();
            int ri = rand.Next(int.MinValue, int.MaxValue);

            if (SSE41Round((float)rd) != ManagedRound((float)rd))
            {
                Console.WriteLine(rd);
            }

            if (SSE41Round((float)(rd * ri)) != ManagedRound((float)(rd * ri)))
            {
                Console.WriteLine(rd);
            }
        }

        Console.WriteLine("Done.");
    }

    [MethodImpl(MethodImplOptions.NoOptimization | MethodImplOptions.NoInlining)]
    static float ManagedRound(float x) => MathF.Round(x);

    [MethodImpl(MethodImplOptions.AggressiveOptimization | MethodImplOptions.NoInlining)]
    static float SSE41Round(float x) => MathF.Round(x);
}
```