@jkotas 2019-07-26
> Also can you briefly educate me how tiering caused htis?

This bug is about observable floating point behavior differences with optimizations on vs. off. Tiering switches between optimization off and on mid-flight that makes this type of bugs to be more severe.

This specific bug is not new. It has been in the product for a long time. The following program will produce difference results for  `dotnet run` vs. `dotnet run -c Release` on .NET Core 2.1:

```
using System;

class Program
{
    static float x = 0.499999969f;

    static void Main(string[] args)
    {
        Console.WriteLine(MathF.Round(x));
    }
}
```

What is our overall confidence that floating point computation produce exact same results with optimizations on vs. optimizations off? If we are not confident in this, "safe fix" may be to disable tiering for any methods with floating point math.