@EgorBo 2019-07-26
@jkotas at least it couldn't happen during the same app session/process at some random point.
I was told hotspot folks had several similar interp/C1/C2 issues with FP methods.

I guess it makes sense to revise all `XX.IsSupported` paths for bugs like this one (because as far as I understand R2R contains only software fallbacks). Or as you said - always tier1 all methods with specific intrinsics inside.