@tannergooding 2019-07-26
> What is our overall confidence that floating point computation produce exact same results with optimizations on vs. optimizations off? If we are not confident in this, "safe fix" may be to disable tiering for any methods with floating point math.

I created and ran the following program for the entire 32-bit floating point range. Aside from NaN differences (software preserves the exact NaN, hardware normalizes to a single NaN), the results are identical: https://gist.github.com/tannergooding/a55001ef69b6ec0746fc05468786cf80