@tannergooding 2019-07-26
(I also found a bug in the codegen for `Sse41.Round*(Vector128<T>)` and will be putting up a fix momentarily; this bug does not impact the 2 argument overloads).