@danmosemsft 2019-07-28
So there are 3 implementations 
1. roundsd and roundss (when not Tier0 if SSE4.1 available)
1. MathF/Math (otherwise)
1. FloatingPointUtils::round (used when evaluating constant expressions during code generation?)