@tannergooding 2019-07-29
> So there are 3 implementations

Correct, although the second two are meant to be identical copies of each other (the former in C# and the latter in C++).