@tannergooding 2019-08-02
Could I get review/sign-off on this for .NET 5 at least, and then we can take it to tactics on Tuesday?

Or are you wanting a different change for release/3.0?