@danmosemsft 2019-08-02
@jkotas is back Monday, I am interested in his take and someone on @dotnet/jit-contrib to help choose correct fix for 3.0 if any.

>  If we are not confident in this, "safe fix" may be to disable tiering for any methods with floating point math.

@kouvel thoughts on the above? It seems this is not a new bug, it is just not encountered easily without tiering.