@tannergooding 2019-08-02
The confidence is fairly high given:
> I created and ran the following program for the entire 32-bit floating point range. Aside from NaN differences (software preserves the exact NaN, hardware normalizes to a single NaN), the results are identical: https://gist.github.com/tannergooding/a55001ef69b6ec0746fc05468786cf80