@danmosemsft 2019-08-02
> The confidence is fairly high given:

Point taken. @dotnet/jit-contrib ? On the numerics side, @GrabYourPitchforks any concern?