@jkotas 2019-08-03
This fix for `Math.Round` looks fine to me. 

> What is our overall confidence that floating point computation produce exact same results with optimizations on vs. optimizations off?

I meant floating point computation in general, not specific to `Math.Round`. For example, https://github.com/dotnet/coreclr/issues/20561 was a similar bug where floating point computation returned different results with optimizations on and off. How many more similar bugs are lurking there? What can we do to find them?